refactor: acceptanceTest 정리
