+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/acceptance/portfolio/PortfolioReadAcceptanceTest.java
@@ -19,7 +19,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 
-class PortfolioAcceptanceTest_Query extends AcceptanceTest {
+class PortfolioReadAcceptanceTest extends AcceptanceTest {
 
     @Autowired
     private PortfolioDtoAssembler portfolioDtoAssembler;
