+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/acceptance/post/PostSearchAcceptanceTest.java
@@ -23,7 +23,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class PostAcceptanceTest_searchPost extends AcceptanceTest {
+class PostSearchAcceptanceTest extends AcceptanceTest {
 
     @BeforeEach
     void setUp() {
