+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/acceptance/tag/TagExtractionAcceptanceTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-class TagAcceptanceTest extends AcceptanceTest {
+class TagExtractionAcceptanceTest extends AcceptanceTest {
 
     @DisplayName("특정 User의 Repository에 기술된 언어 태그들을 추출한다.")
     @Test
