+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/acceptance/user/UserProfileReadAcceptanceTest.java
@@ -8,23 +8,18 @@ import static com.woowacourse.pickgit.common.fixture.TUser.MARK;
 import static com.woowacourse.pickgit.common.fixture.TUser.NEOZAL;
 import static com.woowacourse.pickgit.common.fixture.TUser.모든유저;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.tuple;
 
 import com.woowacourse.pickgit.acceptance.AcceptanceTest;
 import com.woowacourse.pickgit.exception.dto.ApiErrorResponse;
-import com.woowacourse.pickgit.user.application.dto.response.ContributionResponseDto;
-import com.woowacourse.pickgit.user.application.dto.response.UserSearchResponseDto;
 import com.woowacourse.pickgit.user.presentation.dto.response.UserProfileResponse;
-import io.restassured.common.mapper.TypeRef;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-class UserAcceptanceTest_Query extends AcceptanceTest {
+class UserProfileReadAcceptanceTest extends AcceptanceTest {
 
     @BeforeEach
     void setUp() {
@@ -110,74 +105,6 @@ class UserAcceptanceTest_Query extends AcceptanceTest {
     @Test
     void getUserProfile_GuestUser_400Exception() {
         ExtractableResponse<Response> extractableResponse = GUEST.는().프로필을_조회한다(GUEST);
-        assertThat(extractableResponse.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-
-        ApiErrorResponse response = extractableResponse.as(ApiErrorResponse.class);
-        assertThat(response.getErrorCode()).isEqualTo("U0001");
-    }
-
-    @DisplayName("로그인 - 저장된 유저중 유사한 이름을 가진 유저를 검색할 수 있다. 단, 자기 자신은 검색되지 않는다.(팔로잉 여부 true/false)")
-    @Test
-    void searchUser_LoginUser_Success() {
-        List<UserSearchResponseDto> response = KEVIN.은로그인을하고().유저를_검색한다("K")
-            .as(new TypeRef<>() {
-            });
-
-        assertThat(response)
-            .hasSize(2)
-            .extracting("username", "following")
-            .containsExactlyInAnyOrder(
-                tuple(MARK.name(), false),
-                tuple(KODA.name(), false)
-            );
-    }
-
-    @DisplayName("비 로그인 - 저장된 유저중 유사한 이름을 가진 유저를 검색할 수 있다. (팔로잉 필드 null)")
-    @Test
-    void searchUser_GuestUser_Success() {
-        List<UserSearchResponseDto> response = GUEST.는().유저를_검색한다("K")
-            .as(new TypeRef<>() {
-            });
-
-        assertThat(response)
-            .hasSize(3)
-            .extracting("username", "following")
-            .containsExactlyInAnyOrder(
-                tuple(MARK.name(), null),
-                tuple(KEVIN.name(), null),
-                tuple(KODA.name(), null)
-            );
-    }
-
-    @DisplayName("사용자는 활동 통계를 조회할 수 있다.")
-    @Test
-    void getContributions_LoginUser_Success() {
-        ContributionResponseDto response = NEOZAL.은로그인을하고().활동_통계를_조회한다(MARK)
-            .as(ContributionResponseDto.class);
-
-        assertThat(response)
-            .usingRecursiveComparison()
-            .isEqualTo(ContributionResponseDto.builder()
-                .starsCount(11)
-                .commitsCount(48)
-                .prsCount(48)
-                .issuesCount(48)
-                .reposCount(48)
-                .build());
-    }
-
-    @DisplayName("유효하지 않은 토큰으로 활동 통계를 조회할 수 없다. - 401 예외")
-    @Test
-    void getContributions_invalidToken_401Exception() {
-        ApiErrorResponse response = GUEST.는().비정상토큰으로_활통통계를_조회한다(NEOZAL)
-            .as(ApiErrorResponse.class);
-        assertThat(response.getErrorCode()).isEqualTo("A0001");
-    }
-
-    @DisplayName("유효하지 않은 유저 이름으로 활동 통계를 조회할 수 없다. - 400 예외")
-    @Test
-    void getContributions_invalidUsername_400Exception() {
-        ExtractableResponse<Response> extractableResponse = NEOZAL.은로그인을하고().활동_통계를_조회한다(GUEST);
         assertThat(extractableResponse.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
 
         ApiErrorResponse response = extractableResponse.as(ApiErrorResponse.class);
